AllNft calls & technical debt (#731)
* Wallet Exception

Wallet Exception

* Webgl template check

* Published Solution Dependencies to Package Libraries as DLLs

* all erc model

* Auto-duplicate Packages Samples

* complete

complete

* Auto-duplicate Packages Samples

* remove all nft calls

remove all nft calls

* Auto-duplicate Packages Samples

---------
