--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Model/CreateRedeemVoucherModel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 943031bef818de949bb57a89d5733442
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
