--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Model/GetVoucherModel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 19a761f5333299c48935ee82d18c6d3a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
