--- Packages/io.chainsafe.web3-unity/Runtime/Scripts/Model/RedeemVoucherTxModel.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ae75ff6dc5a5ce64fba469f4bb628823
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
